% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akids.R
\name{a_KIDS}
\alias{a_KIDS}
\title{a_KIDS}
\usage{
a_KIDS(x, y, delta, n1, d, fdr, swap = F)
}
\arguments{
\item{x}{a matrix of covariates}

\item{y}{observed time points}

\item{delta}{the censoring indicator where 1 indicates the event, 0 indicates censored.}

\item{n1}{the desired sample size for the screening step
Note: n1+n2=n and n1<n2}

\item{d}{number of predictors to be selected in the screening step}

\item{fdr}{a single/vector of prespecified type I error rate.}

\item{swap}{if 'TRUE', interchanges two continuous variable (X and Y) while calculating the smoothing Kernel G().
default is 'FALSE'.}
}
\value{
a list of active predictors while controlling the prespecified false discovery rate (FDR)
}
\description{
a_KIDS-alpha-controlled kernel-based independence dual screening
}
\author{
Atika Farzana Urmi, Chenlu Ke
}
