% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKIDS_function.R
\name{CKIDS}
\alias{CKIDS}
\title{CKIDS}
\usage{
CKIDS(x, y, delta, z, swap)
}
\arguments{
\item{x}{a matrix of covariates}

\item{y}{observed time points}

\item{delta}{the censoring indicator where 1 indicates the event, 0 indicates censored.}

\item{z}{conditional covariate (categorized)}

\item{swap}{indicates whether inverse regression method should be considered. If swap='TRUE',it will estimate Y given X. otherwise, it will estimate X|Y.}
}
\value{
a list of a matrix of correlation(omega) and the index of covariates based on the highest correlations with outcome conditioning on Z.
}
\description{
CKIDS-Conditional kernel-based independence dual screening
}
\author{
Atika Urmi
}
