% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_d_n2.R
\name{copy_d_n2}
\alias{copy_d_n2}
\title{copy_d_n2}
\usage{
copy_d_n2(data_d_n2, rand_num)
}
\arguments{
\item{data_d_n2}{A data frame of time,delta and original 'd' covariates}

\item{rand_num}{a random number which was used in the data splitting step}
}
\value{
a data frame of time,delta and knockoff copies of original covariates
}
\description{
creating knockoff copies of covairates screened in 1st step.Here we put the bigger data (n2)
with only the covariates selected in the screening step
}
\author{
Atika Farzana Urmi
}
