% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECCFIC_screen.R
\name{ECCFIC_screen}
\alias{ECCFIC_screen}
\title{ECCFIC-screen}
\usage{
ECCFIC_screen(time, delta, x_mat, kernel = "gaussian")
}
\arguments{
\item{time}{a numeric vector of survival time}

\item{delta}{a numeric vector of censoring indicator}

\item{x_mat}{a matrix/dataframe of continious covariates}

\item{kernel}{a kernel to use for x, 'gaussian' or 'distance',default gaussian.
if for 'gaussian' kernel, bandwidth is defined as the heuristic median pairwise distances of x}
}
\value{
ECCFIC correlation between X and Y

the corrleation between X and time given delta,correlation between x and delta and the rank of the
covariates based on importance
}
\description{
ECCFIC_screen ranks the predictors based on their ECCFIC-correlation with survival time and censoring indicator
}
\author{
Atika Farzana Urmi
}
