% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECCFIC_function.R
\name{ECH2}
\alias{ECH2}
\title{ECCFIC- correlation}
\usage{
ECH2(
  x,
  y,
  est = 1,
  kernel = "gaussian",
  sigma = "default",
  alpha = 1,
  bw = "default"
)
}
\arguments{
\item{x}{a numeric vector/matrix}

\item{y}{a numeric vector}

\item{est}{1 for continuous y, 2 for categorical y}

\item{kernel}{a kernel to use for x, 'gaussian' or 'distance'}

\item{sigma}{bandwidth, if a 'gaussian' kernel is used,
default is heuristic median pairwise distances of x}

\item{alpha}{exponent on distance in (0,2], if a 'distance' kernel is used}

\item{bw}{bandwidth of the gaussian smoothing kernel applied on y,
bandwidths suggested by Silverman (1986) are used unless otherwise specified.}
}
\value{
ECCFIC correlation between X and Y

ECCFIC correlation between X and Y
}
\description{
This function calculates the ECCFIC based correlation between two random vector X and Y.
X is always continious here, Y can be either continuous or discrete
}
\author{
Chenlu Ke
}
