% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESFDRC_func.R
\name{ESFDRC_func}
\alias{ESFDRC_func}
\title{ESFDRC}
\usage{
ESFDRC_func(
  data_n1,
  data_n2,
  rand_num,
  q,
  s = round(nrow(data_n1)/log(nrow(data_n1)), 0)
)
}
\arguments{
\item{data_n1}{a data set with column 1= 'time',column2='delta' and rest are 'covariates' (can be obtained from the function split_data())}

\item{data_n2}{a data set with column 1= 'time',column2='delta' and rest are 'covariates' (can be obtained from the function split_data())
Note: n1<n2}

\item{rand_num}{a random seed to reproduce the result}

\item{q}{a prespecified false discovery rate (usually .05 or .10)}

\item{s}{the number of covariates to be screened in the 1st step. Default is (n/log(n))
following Fan&Lv(2008) where 'n' i the number of rows in 'data_n1'}
}
\value{
The final set of important covariates with a specfic false discovery rate
}
\description{
ECCFIC-Screening with False Discovery Rate control
}
\author{
Atika Farzana Urmi
}
