% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{G}
\alias{G}
\title{Smoothing Kernel G(.)}
\usage{
G(y, n)
}
\arguments{
\item{n}{length of x}

\item{x}{the vector to calculate the kernel}
}
\value{
Smoothing Kernel (G) for a vector
}
\description{
Smoothing Kernel G(.)
}
\author{
Chenlu Ke
}
