% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_n1.R
\name{index_n1}
\alias{index_n1}
\title{A sub sample of size n1 with censoring rate similar to original data}
\usage{
index_n1(delta, n1)
}
\arguments{
\item{delta}{a vector of censoring indicator in the original data. This is required to calculate the censoring rate in original data.}

\item{n1}{desired sample size in the sub data}
}
\value{
An index of selected samples of size n1.
}
\description{
A subdata of size n1 that mimics the proportion of censored observations in the original data.
}
\author{
Atika Urmi
}
