% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_n1z.R
\name{index_n1z}
\alias{index_n1z}
\title{A sub sample of size n1 with similar proportion of different levels of delta and Z as in the original data.}
\usage{
index_n1z(delta, n1, z)
}
\arguments{
\item{delta}{a vector of censoring indicator in the original data.}

\item{n1}{desired sample size in the sub data}

\item{Z}{a vector of conditional covariate (categorized)}
}
\value{
An index of selected samples of size n1
}
\description{
A subdata of size n1 that mimics the proportion of different levels of delta and Z as in the original data.
}
\author{
Atika Urmi
}
