% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KIDS_function.R
\name{KIDS}
\alias{KIDS}
\title{KIDS}
\usage{
KIDS(x, y, delta, swap = F)
}
\arguments{
\item{x}{a matrix of covariates}

\item{y}{observed time points}

\item{delta}{the censoring indicator where 1 indicates the event, 0 indicates censored.}

\item{swap}{if 'TRUE', interchanges two continuous variable (X and Y) while calculating the smoothing Kernel G().
default is 'FALSE'.}
}
\value{
a list of a matrix of correlation(omega) and the index of covariates based on the highest correlations.
}
\description{
KIDS-kernel-based independence dual screening
}
\author{
Chenlu Ke, Atika Urmi
}
