% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NW weighting function.R
\name{NW_RBF}
\alias{NW_RBF}
\title{Nadara-Watson estimates of smoothing kernel G}
\usage{
NW_RBF(y, n)
}
\arguments{
\item{y}{the one dimensional vector for which the density is to be estimated}

\item{n}{length of y}
}
\value{
returns a (nxn) matrix of the estimated probability density function (PDF) of a random vector
}
\description{
Nadara-Watson estimates of smoothing kernel G
}
\author{
Chenlu Ke
}
