% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split_data}
\usage{
split_data(data, n1, n2, rand_num)
}
\arguments{
\item{data}{the whole dataset to be splitted into two subsets}

\item{n1}{number of rows in 1st subset}

\item{n2}{number of rows in 2nd subset}

\item{rand_num}{setting a random number to reproduce the result}
}
\value{
Splits the data into two subsets (n1 & n2 with n1<n2) to perform screening with FDR control

A list od two data sets
}
\description{
ECCFIC_screen ranks the predictors based on their ECCFIC-correlation with survival time and censoring indicator
}
\author{
Atika Farzana Urmi
}
