% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{threshold for aKIDS}
\usage{
threshold(W1, W2, fdr)
}
\arguments{
\item{W1}{difference in omega1 between x (original) and xk (kncokoff) for the d covariates selected in screening step}

\item{W2}{difference in omega2 between x (original) and xk (kncokoff) for the d covariates selected in screening step
Note: n1<n2}

\item{fdr}{a single or vector of prespecified false discovery rate for which the threshold to be calculated}
}
\value{
a value for a prespecified FDR to seperate active and inactive set of predictors in aKIDS procedure
}
\description{
This function returns the threshold value for the FDR control part based on the given important statistics (W1,W2)
}
\author{
Chenlu Ke, Atika Urmi
}
