% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_CKIDS.R
\name{a_CKIDS}
\alias{a_CKIDS}
\title{a_CKIDS}
\usage{
a_CKIDS(x, y, delta, n1, d, z, fdr, swap)
}
\arguments{
\item{x}{a matrix of covariates}

\item{y}{observed time points}

\item{delta}{the censoring indicator where 1 indicates the event, 0 indicates censored.}

\item{n1}{the desired sample size for the screening step
Note: n1+n2=n and n1<n2}

\item{d}{number of predictors to be selected in the screening step}

\item{z}{conditional covariate (categorized)}

\item{fdr}{a single/vector of prespecified type I error rate.}

\item{swap}{indicates whether inverse regression method should be considered. If swap='TRUE',it will estimate Y given X. otherwise, it will estimate X|Y.}
}
\value{
lists containing covariates selected in the screening step (d1) and final step (d2) conditioning on Z with a given FDR level
}
\description{
aCKIDS-alpha controlled Conditional kernel-based independence dual screening
}
\author{
Atika Urmi
}
