% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_n1.R
\name{index_n1}
\alias{index_n1}
\title{A sub sample of size n1 with censoring rate similar to original data}
\usage{
index_n1(delta, n1)
}
\arguments{
\item{delta}{a vector of censoring indicator in the original data. This is required to calculate the censoring rate in original data.}

\item{n1}{desired sample size in the sub data}
}
\value{
An index of of size n1 with with desired censoring proportion.
}
\description{
For a data with size n and specific censoring rate, this function returns the index of a sub data of desired size n1 and similar censoring proportion.
}
\author{
Atika Urmi
}
