% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NW kernel regression.R
\name{NW_RBF}
\alias{NW_RBF}
\title{Nadaraya-Watson kernel regression}
\usage{
NW_RBF(y, n)
}
\arguments{
\item{y}{the vector to calculate the kernel}

\item{n}{length of y}
}
\value{
returns a (nxn) matrix of kernel regression estimate using Nadaraya-Watson kernel regression
}
\description{
Nadaraya-Watson kernel regression
}
\author{
Chenlu Ke
}
