% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split_data}
\usage{
split_data(data, n1, n2, rand_num)
}
\arguments{
\item{data}{the whole dataset to be splitted into two subsets. The 1st column has to have the name 'time'
and the 2nd column has to have the name 'delta'. Rest will be the covariates.}

\item{n1}{desired number of rows in 1st subset.
Final n1 might differ slightly after adjusting for censoring rate in both data.}

\item{n2}{desired number of rows in 2nd subset.
Final n2 might differ slightly after adjusting for censoring rate in both data.}

\item{rand_num}{setting a random number to reproduce the result}
}
\value{
Splits the data into two subsets (n1 & n2 with n1<n2) to perform screening with FDR control

A list od two data sets
}
\description{
ECCFIC_screen ranks the predictors based on their ECCFIC-correlation with survival time and censoring indicator
}
\author{
Atika Farzana Urmi
}
